refactor: package 구조 수정 - WebServer, Application, Database 구조가 명확해 지도록 수정합니다.
