+++ src/main/java/codesquad/webserver/WebServer.java
@@ -1,8 +1,5 @@
-package codesquad;
+package codesquad.webserver;
 
-import codesquad.webserver.HttpTask;
-import codesquad.webserver.HttpThreadPool;
-import codesquad.webserver.RequestHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
