+++ src/test/java/codesquad/util/Base64UtilTest.java
@@ -1,5 +1,6 @@
 package codesquad.util;
 
+import codesquad.webserver.util.Base64Util;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
