+++ src/test/java/codesquad/util/StringUtilTest.java
@@ -1,5 +1,6 @@
 package codesquad.util;
 
+import codesquad.webserver.util.StringUtil;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
